+++ 1 The Physical/Guess Point Level 1 (Stateless)/2 - Arrange-Act-Assert Backwards/2.1 - Password Validator/src/password-validator.ts
@@ -1,0 +1,42 @@
+type PasswordValidation = { status: boolean, errors: Array<string> };
+
+function hasUppercase(password: string) {
+    const uppercasePatternRegex: RegExp = /.*[A-Z].*/g;
+    return password.match(uppercasePatternRegex) === null;
+}
+
+function hasMinLength(password: string) {
+    const MIN_LENGTH_FOR_A_PASSWORD = 5;
+    return password.length < MIN_LENGTH_FOR_A_PASSWORD;
+}
+
+function hasMaxLength(password: string) {
+    const MAX_LENGTH_FOR_A_PASSWORD = 15;
+    return password.length > MAX_LENGTH_FOR_A_PASSWORD;
+}
+
+function hasAtLeastANumber(password: string) {
+    const atLeastANumberPatternRegex: RegExp = /.*[0-9]+.*/g;
+    return password.match(atLeastANumberPatternRegex) === null;
+}
+
+export function passwordValidator(password: string): PasswordValidation {
+
+    if (hasAtLeastANumber(password)) {
+        return { status: false, errors: ['NO_DIGITS']};
+    }
+
+    if (hasMaxLength(password)) {
+        return { status: false, errors: ['TOO_LONG']};
+    }
+
+    if (hasMinLength(password)) {
+        return { status: false, errors: ['TOO_SHORT']};
+    }
+
+    if (hasUppercase(password)) {
+        return { status: false, errors: ['NO_UPPER']};
+    }
+
+    return { status: true, errors: []};
+}
